Move all account code from api to accounts app
