--- project/api/authentication.py
@@ -1,7 +1,0 @@
-from rest_framework import authentication
-
-
-class CsrfExemptSessionAuthentication(authentication.SessionAuthentication):
-    """ """
-    def enforce_csrf(self, request):
-        return
