+++ project/api/read.py
@@ -1,125 +1,9 @@
-from django.contrib.auth import get_user_model
 from django.http import JsonResponse, HttpResponseBadRequest
 from django.forms.models import model_to_dict
 
 from .models import Thread, Civi, Activity
 from accounts.models import Profile
 from .utils import json_response
-
-
-def get_user(request, user):
-    """
-    USAGE:
-        This is used to get a user
-    """
-
-    User = get_user_model()
-
-    try:
-        u = User.objects.get(username=user)
-        a = Profile.objects.get(user=u)
-
-        return JsonResponse(model_to_dict(a))
-    except Profile.DoesNotExist as e:
-        return HttpResponseBadRequest(reason=str(e))
-
-
-def get_card(request, user):
-    """
-    USAGE:
-        This is used to get a card
-    """
-
-    User = get_user_model()
-
-    try:
-        u = User.objects.get(username=user)
-        a = Profile.objects.get(user=u)
-        result = Profile.objects.card_summarize(
-            a, Profile.objects.get(user=request.user)
-        )
-        return JsonResponse(result)
-    except User.DoesNotExist:
-        return HttpResponseBadRequest(reason=f"User with username {user} not found")
-    except Profile.DoesNotExist as e:
-        return HttpResponseBadRequest(reason=str(e))
-    except Exception as e:
-        return HttpResponseBadRequest(reason=str(e))
-
-
-def get_profile(request, user):
-    """
-    USAGE:
-       This is used to get a user profile
-    """
-
-    User = get_user_model()
-
-    try:
-        u = User.objects.get(username=user)
-        a = Profile.objects.get(user=u)
-        result = Profile.objects.summarize(a)
-
-        result["issues"] = []
-        voted_solutions = Activity.objects.filter(
-            account=a.id, civi__c_type="solution", activity_type__contains="pos"
-        )
-
-        solution_threads = voted_solutions.values("thread__id").distinct()
-
-        for thread_id in solution_threads:
-            t = Thread.objects.get(id=thread_id)
-            solutions = []
-            solution_civis = voted_solutions.filter(thread=thread_id).values_list(
-                "civi__id", flat=True
-            )
-            for civi_id in solution_civis:
-                c = Civi.objects.get(id=civi_id)
-                vote = voted_solutions.get(civi__id=civi_id).activity_type
-                vote_types = {"vote_pos": "Agree", "vote_vpos": "Strongly Agree"}
-                solution_item = {
-                    "id": c.id,
-                    "title": c.title,
-                    "body": c.body,
-                    "user_vote": vote_types.get(vote),
-                }
-                solutions.append(solution_item)
-
-            my_issue_item = {
-                "thread_id": t.id,
-                "thread_title": t.title,
-                "category": t.category.name,
-                "solutions": solutions,
-            }
-            result["issues"].append(my_issue_item)
-
-        if request.user.username != user:
-            ra = Profile.objects.get(user=request.user)
-            if user in ra.following.all():
-                result["follow_state"] = True
-            else:
-                result["follow_state"] = False
-
-        return JsonResponse(result)
-
-    except Profile.DoesNotExist as e:
-        return HttpResponseBadRequest(reason=str(e))
-
-
-def get_feed(request):
-    """
-    USAGE:
-       This is used to get a feed for a user
-    """
-    try:
-        feed_threads = [
-            Thread.objects.summarize(t) for t in Thread.objects.order_by("-created")
-        ]
-
-        return json_response(feed_threads)
-
-    except Exception as e:
-        return HttpResponseBadRequest(reason=str(e))
 
 
 def get_thread(request, thread_id):
